double, int, and nsnumber localization methods added
