+++ AKLanguageManager/Source/Extensions/Bundle+Swizzle.swift
@@ -1,5 +1,5 @@
 //
-//  Bundle+Swizzling.swift
+//  Bundle+Swizzle.swift
 //  AKLanguageManager
 //
 //  Created by Amr Koritem on 13/09/2022.
