+++ AKLanguageManager/Source/Extensions/String+Localized.swift
@@ -1,11 +1,11 @@
 //
-//  Localized+Utilities.swift
+//  String+Localized.swift
 //  AKLanguageManager
 //
 //  Created by Amr Koritem on 13/09/2022.
 //
 
-import UIKit
+import Foundation
 
 public extension String {
     /// Localize the entire string.
@@ -71,14 +71,18 @@ public extension String {
     }
 
     /// Localize the numbers only to the designated language.
-    func numbersLocalized(in language: Language) -> String {
+    /// - Parameters:
+    ///   - language: The language to which the string will be localized.
+    ///   - numberStyle: The style of the returned number localization.
+    /// - Returns:
+    ///   - The string with all its numbers localized in the designated language.
+    func numbersLocalized(in language: Language, style numberStyle: NumberFormatter.Style = .decimal) -> String {
+        let regex = "\(Language.en.numberRegex(minNumberOfDigits: 1))|\(Language.ar.numberRegex(minNumberOfDigits: 1))"
         var localizedString = doublesPreparedForLocalization(in: language)
         let formatter = NumberFormatter()
-        formatter.numberStyle = .decimal
+        formatter.numberStyle = numberStyle
         formatter.locale = language.locale
-        let regex = "\(Language.en.numberRegex(minNumberOfDigits: 1))|\(Language.ar.numberRegex(minNumberOfDigits: 1))"
-        let matches = matchesForRegex(regex).uniqued()
-        matches.forEach { match in
+        matchesForRegex(regex).uniqued().forEach { match in
             guard let nsNumberMatch = formatter.number(from: match),
                   let localizedMatch = formatter.string(from: nsNumberMatch) else { return }
             localizedString = localizedString.replacingOccurrences(of: match, with: localizedMatch)
@@ -97,115 +101,6 @@ public extension String {
             preparedString = preparedString.replacingOccurrences(of: match, with: modifiedMatch)
         }
         return preparedString
-    }
-}
-
-public extension NSAttributedString {
-    /// Localize the entire string while keeping its attributes.
-    var localized: NSAttributedString {
-        let attributedString = NSMutableAttributedString(attributedString: self)
-        attributedString.mutableString.setString(string.localized)
-        return attributedString
-    }
-
-    /// Localize the entire string to the designated language while keeping its attributes.
-    func localized(in language: Language) -> NSAttributedString {
-        let attributedString = NSMutableAttributedString(attributedString: self)
-        attributedString.mutableString.setString(string.localized(in: language))
-        return attributedString
-    }
-}
-
-extension UIImage {
-    struct AssociatedKeys {
-        static var isRightToLeft: UInt8 = 0
-    }
-
-    /// Check if the image is flipped in right to left direction.
-    @objc
-    public internal(set) var isRightToLeft: Bool {
-        get {
-            let isRightToLeft = objc_getAssociatedObject(self, &AssociatedKeys.isRightToLeft) as? Bool
-            return isRightToLeft ?? false
-        }
-        set {
-            objc_setAssociatedObject(
-                self,
-                &AssociatedKeys.isRightToLeft,
-                newValue,
-                objc_AssociationPolicy.OBJC_ASSOCIATION_RETAIN_NONATOMIC
-            )
-        }
-    }
-
-    /// The current image direction.
-    public var horizontalDirection: Direction {
-        isRightToLeft ? .rightToLeft : .leftToRight
-    }
-
-    /// Returns a version of the image that's flipped in the opposite direction.
-    public var horizontalDirectionReverted: UIImage? {
-        horizontalDirectionChanged(to: horizontalDirection.opposite)
-    }
-
-    /// Returns a version of the image that's flipped in right to left direction or left to right direction depending on the current language.
-    public var directionLocalized: UIImage? {
-        directionLocalized(in: AKLanguageManager.shared.selectedLanguage)
-    }
-
-    /// Determines if the current `imageOrientation` is mirrored.
-    public var isOrientationMirrored: Bool {
-        imageOrientation.rawValue > 3
-    }
-
-    var horizontallyFlipped: UIImage? {
-        guard let cgImage = cgImage else { return nil }
-        // Using just `withHorizontallyFlippedOrientation()` causes sf-symbols to not change their orientation.
-        return UIImage(
-            cgImage: cgImage,
-            scale: scale,
-            orientation: withHorizontallyFlippedOrientation().imageOrientation)
-    }
-
-    /// Returns a version of the image that's flipped in right to left direction or left to right direction depending on the designated language.
-    public func directionLocalized(in language: Language) -> UIImage? {
-        horizontalDirectionChanged(to: language.isRightToLeft ? .rightToLeft : .leftToRight)
-    }
-
-    /// Returns a version of the image that's flipped in the direction specified.
-    public func horizontalDirectionChanged(to direction: Direction) -> UIImage? {
-        let shouldChangeDirection = horizontalDirection != direction
-        guard shouldChangeDirection else { return self }
-        let image = horizontallyFlipped
-        image?.isRightToLeft = direction == .rightToLeft
-        return image
-    }
-
-    public enum Direction {
-        case rightToLeft
-        case leftToRight
-
-        public var opposite: Direction {
-            self == .rightToLeft ? .leftToRight : .rightToLeft
-        }
-    }
-}
-
-public extension NSTextAlignment {
-    /// Returns an `NSTextAlignment` enum suitable for the current language.
-    static var localized: NSTextAlignment {
-        localized(in: AKLanguageManager.shared.selectedLanguage)
-    }
-
-    /// Returns an `NSTextAlignment` enum suitable for the current language.
-    var localized: NSTextAlignment {
-        guard self == .natural else { return self }
-        return .localized
-    }
-
-    /// Returns an `NSTextAlignment` enum suitable for the designated language.
-    static func localized(in language: Language) -> NSTextAlignment {
-        return language.isRightToLeft ? .right : .left
     }
 }
