+++ AKLanguageManager/Source/Extensions/UIView+Swizzle.swift
@@ -1,5 +1,5 @@
 //
-//  UIView+Swizzling.swift
+//  UIView+Swizzle.swift
 //  AKLanguageManager
 //
 //  Created by Amr Koritem on 13/09/2022.
