+++ AKLanguageManagerTests/LocalizedUtilitiesTests.swift
@@ -19,6 +19,42 @@ class LocalizedUtilitiesTests: XCTestCase {
         languageManager.defaultLanguage = .en
     }
 
+    func testIntLocalized() {
+        let int = 12
+        XCTAssertEqual(int.localized, "12")
+        languageManager.setLanguage(language: .ar)
+        XCTAssertEqual(int.localized, "١٢")
+        XCTAssertEqual(int.localized(in: .en), "12")
+        XCTAssertEqual(int.localized(in: Language.en.locale) ?? "", "12")
+    }
+
+    func testDoubleLocalized() {
+        let double = 12.2
+        XCTAssertEqual(double.localized, "12.2")
+        languageManager.setLanguage(language: .ar)
+        XCTAssertEqual(double.localized, "١٢,٢")
+        XCTAssertEqual(double.localized(in: .en), "12.2")
+        XCTAssertEqual(double.localized(in: Language.en.locale) ?? "", "12.2")
+    }
+
+    func testNSNumberDoubleLocalized() {
+        let nsDouble = NSNumber(value: 12.2)
+        XCTAssertEqual(nsDouble.localized, "12.2")
+        languageManager.setLanguage(language: .ar)
+        XCTAssertEqual(nsDouble.localized, "١٢,٢")
+        XCTAssertEqual(nsDouble.localized(in: .en), "12.2")
+        XCTAssertEqual(nsDouble.localized(in: Language.en.locale) ?? "", "12.2")
+    }
+
+    func testNSNumberLocalized() {
+        let nsInt = NSNumber(value: 12)
+        XCTAssertEqual(nsInt.localized, "12")
+        languageManager.setLanguage(language: .ar)
+        XCTAssertEqual(nsInt.localized, "١٢")
+        XCTAssertEqual(nsInt.localized(in: .en), "12")
+        XCTAssertEqual(nsInt.localized(in: Language.en.locale) ?? "", "12")
+    }
+
     func testUIImageLocalized() {
         let image = UIImage(named: "image", in: Language.mainBundle, compatibleWith: nil)
         XCTAssertTrue(image?.isRightToLeft == false)
@@ -47,8 +83,6 @@ class LocalizedUtilitiesTests: XCTestCase {
 
     func testNSTextAlignmentLocalizedIn() {
         XCTAssertEqual(NSTextAlignment.localized(in: .en), .left)
-        XCTAssertEqual(NSTextAlignment.localized(in: .en), .left)
-        XCTAssertEqual(NSTextAlignment.localized(in: .ar), .right)
         XCTAssertEqual(NSTextAlignment.localized(in: .ar), .right)
     }
 }
